Start game restrictions

+++ src/constants.rs
@@ -2,8 +2,8 @@
 pub const MAX_GAME_ID: i32 = 100000;
 pub const TIMER_START_VALUE: i32 = 60000;
 pub const MAX_WORDS_IN_PLAY: usize = 2;
-pub const MIN_WORDS_PER_PLAYER: i32 = 1;
-pub const MAX_WORDS_PER_PLAYER: i32 = 7;
+pub const MIN_WORDS_PER_PLAYER: usize = 1;
+pub const MAX_WORDS_PER_PLAYER: usize = 7;
 
 pub const NEW_PLAYER_EVENT_PREFIX: &str = "new_player:";
 pub const TIMER_UPDATE_EVENT_PREFIX: &str = "timer_update:";
